dev-requirements.txt
pytest-ckan
python-jose==3.3.0